refactor: split ast::pl::fold
Just helping the git file rename detection for
the following extraction of the prql_ast crate.
